fix: move snakemake.io.container into io_container to allow python >3.7 in snakemake.io
