+++ src/snakemake/caching/hash.py
@@ -111,6 +111,7 @@ class ProvenanceHashMap:
 
         # Hash input files that are not generated by other jobs (sorted by hash value).
         omit_storage_content = job.rule.cache and job.rule.cache.omit_storage_content
+
         async def get_checksum(f):
             if f.is_storage and omit_storage_content:
                 return f.storage_object.query
