+++ src/snakemake/io_container.py
